+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## UNRELEASED
+
+### Adds
+
+* New `contain-nested` array option to list mixins that are known to contain nested rules.
+
 ## 1.1.0
 
 ### Adds